♻️[Refactor]#5 - viewModel 리팩토링
